methode Delete fixed and styled

+++ frontend/src/components/profile-artist-page/DeleteFlashButton.jsx
@@ -1,11 +1,17 @@
 import React from 'react';
 import { GiCrossMark } from 'react-icons/gi'
 
-function DeletePhotoButton({ albumId, artistId, dataUpdated }) {
+function DeleteFlashButton({ flashId, artistId, dataUpdated }) {
 
   const handleDelete = async () => {
+
+    const userConfirmed = window.confirm('Êtes-vous sûr de vouloir supprimer cette photo ?');
+    if (!userConfirmed) {
+      return;
+    }
+    
     try {
-      const url = `http://127.0.0.1:8000/user_artist/api/ma-page-artiste/${artistId}/delete-album/${albumId}`;
+      const url = `http://127.0.0.1:8000/user_artist/api/ma-page-artiste/${artistId}/delete-flash/${flashId}`;
       const deleteResponse = await fetch(url, {
         method: 'DELETE',
         headers: {
@@ -15,7 +21,7 @@ function DeletePhotoButton({ albumId, artistId, dataUpdated }) {
       });
 
       if (deleteResponse.ok) {
-        // Refetch the artist data to update photos after deletion
+        // Refetch the artist data to update flashs after deletion
         const response = await fetch(`http://127.0.0.1:8000/user_artist/api/ma-page-artiste/${artistId}/`);
         if (response.ok) {
           const data = await response.json();
@@ -38,4 +44,4 @@ function DeletePhotoButton({ albumId, artistId, dataUpdated }) {
   );
 }
 
-export default DeletePhotoButton;
+export default DeleteFlashButton;
